+++ client/blueprintRestClient_test.go
@@ -16,7 +16,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/rest"
 
-	bpv2 "github.com/cloudogu/k8s-blueprint-lib/v2/api/v2"
+	bpv2 "github.com/cloudogu/k8s-blueprint-lib/v3/api/v3"
 )
 
 var testCtx = context.Background()