+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [#11] **Breaking** blueprint mask have now a union type, where they can or added directly or referenced 
-  - if you have a mask in your blueprint, you have to add it in the union type now
+  - if you have a mask in your blueprint, you have to move it into the union type now
 
 ## [v2.0.1] - 2025-10-23